first feature in portfolio

+++ src/app/components/shared/editable/editable.component.ts
@@ -9,6 +9,7 @@ import { ChangedInputComponent } from '../../dialogs/changed-input/changed-input
 import { CommonModule } from '@angular/common';
 import { FirebaseService } from '../../../services/global/firebase.service';
 import { Firestore, collection } from '@angular/fire/firestore';
+import { Todo } from '../../../interfaces/todo';
 
 @Component({
   selector: 'app-editable',
@@ -19,7 +20,7 @@ import { Firestore, collection } from '@angular/fire/firestore';
 })
 export class EditableComponent implements OnInit{
   @Input() groupedForm !: FormGroup;
-  @Input() item ?: any;
+  @Input() item ?: Todo;
   @Input() listTitle !: string;
   @Input() fireListHeader !: string;
   @Output() closeEdit = new EventEmitter<void>();
@@ -39,24 +40,18 @@ export class EditableComponent implements OnInit{
     }
 
   protected delete():void {
-    this.firebaseService.deleteFireItem(this.fireListHeader, this.item.id);
+    if(this.item?.id) this.firebaseService.deleteFireItem(this.fireListHeader, this.item.id);
     this.utilService.closeThis(this.closeWindow, 'editableClose');
   }
 
   protected save():void {
-    const todo = {
-      title: this.groupedForm.value.title,
-      description: this.groupedForm.value.description,
-      done: this.groupedForm.value.done,
-      read: this.groupedForm.value.read,
-    }
-    console.log(this.item);
+    const todo: Todo = { ...this.groupedForm.value };
     this.utilService.closeThis(this.closeWindow, 'editableClose');
-    if (!this.item) {
+    if (this.item?.id) 
+      this.firebaseService.updateFireItem(this.fireCollection, this.item.id, todo);
+    else {
       const token = this.utilService.generateSimpleToken(10);
       this.firebaseService.setItemToFirebase(this.fireListHeader, token, todo);
-    } else {
-      this.firebaseService.updateFireItem(this.fireCollection, this.item.id, todo);
     }
   }
